fixed comments and failing IT tests

+++ Backend/src/main/java/ca/mcgill/cooperator/dto/CoopDetailsDto.java
@@ -2,7 +2,7 @@ package ca.mcgill.cooperator.dto;
 
 public class CoopDetailsDto {
     private int id;
-    private double payPerHour;
+    private int payPerHour;
     private int hoursPerWeek;
 
     private EmployerContactDto employerContact;
@@ -12,7 +12,7 @@ public class CoopDetailsDto {
     
     public CoopDetailsDto(
             int id,
-            double payPerHour,
+            int payPerHour,
             int hoursPerWeek,
             EmployerContactDto employerContact,
             CoopDto coop) {
@@ -29,11 +29,11 @@ public class CoopDetailsDto {
         return this.id;
     }
 
-    public double getPayPerHour() {
+    public int getPayPerHour() {
         return this.payPerHour;
     }
 
-    public void setPayPerHour(double payPerHour) {
+    public void setPayPerHour(int payPerHour) {
         this.payPerHour = payPerHour;
     }
